requirements.txt
django==5.2.3
python-telegram-bot==13.15
python-dotenv==1.1.1
